move out of the esri feature and contentful folder
