+++ src/containers/sidebars/aoi-sidebar/sidebar-card-content/index.js
@@ -11,8 +11,8 @@ import {
   batchToggleLayers,
 } from 'utils/layer-manager-utils';
 
-import ContentfulService from 'services/contentful/contentful';
-import EsriFeatureService from 'services/esri-feature/esri-feature-service';
+import ContentfulService from 'services/contentful';
+import EsriFeatureService from 'services/esri-feature-service';
 
 import {
   AOI_LEGEND_CATEGORIES,
