+++ src/containers/sidebars/map-layers/carbon-sidebar-card/carbon-sidebar-card.js
@@ -8,7 +8,7 @@ import * as urlActions from 'actions/url-actions';
 
 import { layerManagerToggle } from 'utils/layer-manager-utils';
 
-import ContentfulService from 'services/contentful/contentful';
+import ContentfulService from 'services/contentful';
 
 import { CARBON_LAYER } from 'constants/layers-slugs';
 import metadataConfig from 'constants/metadata';
