+++ src/containers/sidebars/national-report-sidebar/overview-sidebar/local-priority-card/local-priority-card.js
@@ -8,7 +8,7 @@ import * as urlActions from 'actions/url-actions';
 
 import { layerManagerToggle } from 'utils/layer-manager-utils';
 
-import ContentfulService from 'services/contentful/contentful';
+import ContentfulService from 'services/contentful';
 
 import metadataConfig, {
   COUNTRY_PRIORITY,
