+++ src/pages/nrc-legacy/index.js
@@ -6,7 +6,7 @@ import * as urlActions from 'actions/url-actions';
 
 import { activateLayersOnLoad, setBasemap } from 'utils/layer-manager-utils';
 
-import EsriFeatureService from 'services/esri-feature/esri-feature-service';
+import EsriFeatureService from 'services/esri-feature-service';
 
 import {
   FIREFLY_BASEMAP_LAYER,
