+++ src/utils/geo-processing-services.js
@@ -1,6 +1,6 @@
 import { loadModules } from 'esri-loader';
 
-import EsriFeatureService from 'services/esri-feature/esri-feature-service';
+import EsriFeatureService from 'services/esri-feature-service';
 import { getCrfData } from 'services/geo-processing-services/biodiversity/biodiversity';
 import { getCrfData as getContextualData } from 'services/geo-processing-services/contextual/contextual-data';
